+++ tools/apiview/emitters/typespec-apiview/src/apiview.ts
@@ -1,6 +1,7 @@
 import {
   AliasStatementNode,
   ArrayExpressionNode,
+  ArrayLiteralNode,
   AugmentDecoratorStatementNode,
   BaseNode,
   BooleanLiteralNode,
@@ -23,6 +24,9 @@ import {
   Namespace,
   navigateProgram,
   NumericLiteralNode,
+  ObjectLiteralNode,
+  ObjectLiteralPropertyNode,
+  ObjectLiteralSpreadPropertyNode,
   OperationSignatureDeclarationNode,
   OperationSignatureReferenceNode,
   OperationStatementNode,
@@ -379,6 +383,7 @@ export class ApiView {
 
   tokenize(node: BaseNode) {
     let obj;
+    let last = 0;  // track the final index of an array
     let isExpanded = false;
     switch (node.kind) {
       case SyntaxKind.AliasStatement:
@@ -395,6 +400,18 @@ export class ApiView {
         obj = node as ArrayExpressionNode;
         this.tokenize(obj.elementType);
         this.punctuation("[]");
+        break;
+      case SyntaxKind.ArrayLiteral:
+        obj = node as ArrayLiteralNode;
+        this.punctuation("#[");
+        last = obj.values.length - 1;
+        obj.values.forEach((val, i) => {
+          this.tokenize(val);
+          if (i !== last) {
+            this.punctuation(",", false, true);
+          }
+        });
+        this.punctuation("]");
         break;
       case SyntaxKind.AugmentDecoratorStatement:
         obj = node as AugmentDecoratorStatementNode;
@@ -435,7 +452,7 @@ export class ApiView {
         this.tokenizeIdentifier(obj.target, "keyword");
         this.lineMarker();
         if (obj.arguments.length) {
-          const last = obj.arguments.length - 1;
+          last = obj.arguments.length - 1;
           this.punctuation("(", false, false);
           for (let x = 0; x < obj.arguments.length; x++) {
             const arg = obj.arguments[x];
@@ -544,6 +561,28 @@ export class ApiView {
         obj = node as NumericLiteralNode;
         this.literal(obj.value.toString());
         break;
+      case SyntaxKind.ObjectLiteral:
+        obj = node as ObjectLiteralNode;
+        this.punctuation("#{", true, false);
+        last = obj.properties.length - 1;
+        obj.properties.forEach((prop, i) => {
+          this.tokenize(prop);
+          if (i !== last) {
+            this.punctuation(",", false, true);
+          }
+        });
+        this.punctuation("}", false, false);
+        break;
+      case SyntaxKind.ObjectLiteralProperty:
+        obj = node as ObjectLiteralPropertyNode;
+        this.tokenizeIdentifier(obj.id, "member");
+        this.punctuation(":", false, true);
+        this.tokenize(obj.value);
+        break;
+      case SyntaxKind.ObjectLiteralSpreadProperty:
+        obj = node as ObjectLiteralSpreadPropertyNode;
+        // TODO: Whenever there is an example?
+        throw new Error(`Case "ObjectLiteralSpreadProperty" not implemented`);
       case SyntaxKind.OperationStatement:
         this.tokenizeOperationStatement(node as OperationStatementNode);
         break;
@@ -641,7 +680,7 @@ export class ApiView {
         this.tokenizeUnionVariant(node as UnionVariantNode);
         break;
       case SyntaxKind.UnknownKeyword:
-        this.keyword("any", true, true);
+        this.keyword("unknown", true, true);
         break;
       case SyntaxKind.UsingStatement:
         throw new Error(`Case "UsingStatement" not implemented`);
@@ -741,7 +780,6 @@ export class ApiView {
           case SyntaxKind.MemberExpression:
             return this.getFullyQualifiedIdentifier(obj.target as MemberExpressionNode);
         }
-        break;
       default:
         throw new Error(`Unsupported expression kind: ${SyntaxKind[node.kind]}`);
       //unsupported ArrayExpressionNode | MemberExpressionNode | ModelExpressionNode | TupleExpressionNode | UnionExpressionNode | IntersectionExpressionNode | TypeReferenceNode | ValueOfExpressionNode | AnyKeywordNode;