+++ tools/apiview/emitters/typespec-apiview/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History
 
+## Version 0.4.9 (07-09-2024)
+Fix issue where "unknown" was rendered as "any".
+Support value syntax for objects and arrays.
+
 ## Version 0.4.8 (04-18-2024)
 Display suppressions in APIView.
 Resolve visual anomalies.